requirements.txt
pytz~=2024.1
aiojobs~=1.3.0
python-dotenv~=1.0.1
gspread~=6.1.2
oauth2client~=4.1.3
pandas~=2.2.2
python-telegram-bot
aiogram~=3.12.0
pillow~=10.4.0